+++ src/plugin/accessories/StationAccessory.ts
@@ -16,6 +16,7 @@ export class StationAccessory {
   private station: Station;
 
   private service: Service;
+  private manualTriggerService: Service;
   private alarm_triggered: boolean;
   private modes;
 
@@ -85,6 +86,20 @@ export class StationAccessory {
       .onGet(this.handleSecuritySystemTargetStateGet.bind(this))
       .onSet(this.handleSecuritySystemTargetStateSet.bind(this));
 
+    this.manualTriggerService = 
+      this.accessory.getService(this.platform.Service.Switch) ||
+      this.accessory.addService(this.platform.Service.Switch);
+    
+    this.manualTriggerService.setCharacteristic(
+      this.characteristic.Name,
+      accessory.displayName + ' alarm switch',
+    );
+
+    this.manualTriggerService
+      .getCharacteristic(this.characteristic.On)
+      .onGet(this.handleManualTriggerSwitchStateGet.bind(this))
+      .onSet(this.handleManualTriggerSwitchStateSet.bind(this));
+
     this.eufyStation.on('guard mode', (station: Station, guardMode: number) =>
       this.onStationGuardModePushNotification(station, guardMode),
     );
@@ -130,6 +145,14 @@ export class StationAccessory {
       config.hkOff = config.hkOff ??= this.platform.config.hkOff;
     }
 
+    if (!Array.isArray(config.manualTriggerModes)) {
+      config.manualTriggerModes = [];
+    }
+    this.platform.log.debug(
+      this.accessory.displayName, 'manual alarm will be triggered only in these hk modes: ' + config.manualTriggerModes);
+
+    config.manualAlarmSeconds = config.manualAlarmSeconds ??= 30;
+
     return config;
   }
 
@@ -165,6 +188,14 @@ export class StationAccessory {
     station: Station,
     alarmEvent: AlarmEvent,
   ): void {
+    let currentValue = this.eufyStation.getPropertyValue(PropertyName.StationCurrentMode);
+    if (alarmEvent === 0) {
+      // do not resset alarm if alarm was triggered manually
+      // since the alarm can only be triggered for 30 seconds for now (limitation of eufy-security-client)
+      // this would mean that the alarm is always reset after 30 seconds
+      // see here: https://github.com/bropat/eufy-security-client/issues/178
+      currentValue = -1;
+    }
     switch (alarmEvent) {
       case 2: // Alarm triggered by GSENSOR
       case 3: // Alarm triggered by PIR
@@ -179,11 +210,17 @@ export class StationAccessory {
           .getCharacteristic(this.characteristic.SecuritySystemCurrentState)
           .updateValue(this.characteristic.SecuritySystemCurrentState.ALARM_TRIGGERED); // Alarm !!!
         break;
+      case 0:  // Alarm off by Hub
       case 15: // Alarm off by Keypad
       case 16: // Alarm off by Eufy App
       case 17: // Alarm off by HomeBase button
         this.platform.log.warn('ON StationAlarmEvent - ALARM OFF - alarmEvent:', alarmEvent);
         this.alarm_triggered = false;
+        if (currentValue !== -1) {
+          this.service
+            .getCharacteristic(this.characteristic.SecuritySystemCurrentState)
+            .updateValue(this.convertEufytoHK(currentValue)); // reset alarm state
+        }
         break;
       default:
         this.platform.log.warn('ON StationAlarmEvent - ALARM UNKNOWN - alarmEvent:', alarmEvent);
@@ -192,6 +229,9 @@ export class StationAccessory {
           .updateValue(this.characteristic.StatusFault.GENERAL_FAULT);
         break;
     }
+    this.manualTriggerService
+      .getCharacteristic(this.characteristic.On)
+      .updateValue(this.alarm_triggered);
   }
 
   private mappingHKEufy(): void {
@@ -259,7 +299,7 @@ export class StationAccessory {
   /**
    * Handle requests to get the current value of the 'Security System Target State' characteristic
    */
-  handleSecuritySystemTargetStateGet(): CharacteristicValue {
+  private handleSecuritySystemTargetStateGet(): CharacteristicValue {
     try {
       const currentValue = this.eufyStation.getPropertyValue(PropertyName.StationCurrentMode);
       if (currentValue === -1) {
@@ -298,8 +338,48 @@ export class StationAccessory {
           this.platform.log.error(this.accessory.displayName, 'Changing guard mode to ' + this.getGuardModeName(value) + ' timed out!');
         }, 5000);
       }, 5000);
+
+      this.manualTriggerService
+        .getCharacteristic(this.characteristic.On)
+        .updateValue(false);
     } catch (error) {
       this.platform.log.error(this.accessory.displayName + ': Error Setting security mode!', error);
+    }
+  }
+
+  private handleManualTriggerSwitchStateGet(): CharacteristicValue {
+    return this.alarm_triggered;
+  }
+
+  private async handleManualTriggerSwitchStateSet(value: CharacteristicValue) {
+    if (value) { // trigger alarm if allowed
+      try {
+        const currentValue = this.eufyStation.getPropertyValue(PropertyName.StationCurrentMode);
+        if (currentValue === -1) {
+          throw 'Something wrong with this device';
+        }
+        if (this.stationConfig.manualTriggerModes.indexOf(this.convertEufytoHK(currentValue)) !== -1) {
+          this.eufyStation.triggerStationAlarmSound(this.stationConfig.manualAlarmSeconds)
+            .then(() => this.platform.log.debug(
+              this.accessory.displayName, 'alarm manually triggered for ' + this.stationConfig.manualAlarmSeconds + ' seconds.'))
+            .catch(err => this.platform.log.error(this.accessory.displayName, 'alarm could not be manually triggered: ' + err));
+        } else {
+          setTimeout(() => {
+            // eslint-disable-next-line max-len
+            this.platform.log.info(this.accessory.displayName, 'tried to trigger alarm, but the current station mode prevents the alarm from being triggered. Please look in in the configuration if you want to change this behaviour.');
+            this.manualTriggerService
+              .getCharacteristic(this.characteristic.On)
+              .updateValue(false);
+          }, 1000);
+        }
+      } catch {
+        this.platform.log.error(this.accessory.displayName, 'handleSecuritySystemTargetStateGet', 'Wrong return value');
+        return;
+      }
+    } else { // reset alarm
+      this.eufyStation.resetStationAlarmSound()
+        .then(() => this.platform.log.debug(this.accessory.displayName, 'alarm manually reset'))
+        .catch(err => this.platform.log.error(this.accessory.displayName, 'alarm could not be reset: ' + err));
     }
   }